Using the new Powerline now for tmux
Just getting started with it. But it looks really nice,
and I think it will stick.
Probably will move information from the prompt to the
powerline over time.

This is currently not installed. The process included
installing python, force overwriting the links and
then installing via pip. Need to add this feature
to exogenesis, see moonglum/exogenesis#12