Use `yakshed/moonglum` :rocket:
This is an immense speed up for a `vagrant up` after a `vagrant
destroy`. I'm talking about 15-20 minutes in the old version and under 1
minute in the new version.

The trick is that almost everything that was done in the provision task
is now done on Atlas with packer. The project can be found here:

https://github.com/yakshed/base-vm

Thanks @railsbros-dirk for pairing with me on that :heart: